style(mac): satisfy Swift CI checks
